requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
boto3==1.28.29
    # via -r requirements.in
botocore==1.31.29
    # via
    #   boto3
    #   s3transfer
build==0.10.0
    # via pip-tools
certifi==2023.7.22
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.2.0
    # via requests
click==8.1.7
    # via pip-tools
idna==3.4
    # via requests
iniconfig==2.0.0
    # via pytest
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
packaging==23.1
    # via
    #   build
    #   pytest
pip-tools==7.3.0
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   botocore
    #   time-machine
python-dotenv==1.0.0
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
s3transfer==0.6.2
    # via boto3
sentry-sdk==1.29.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
time-machine==2.13.0
    # via -r requirements.in
urllib3==1.26.16
    # via
    #   botocore
    #   requests
    #   sentry-sdk
wheel==0.41.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
